Changes: Hardwire the QStyle to use for the Passthrough window
RevBy: Pekka Vuorela
Details: There are issues with the QtMaemo6Style (which is used when libdui is
compiled with plainqt support, as the Passthrough window is treated as a plain
Qt app in this context), e.g., NB #162495. This commit hardwires the QStyle,
disabling QtMaemo6Style explicitly. This change should be revised at a later
stage (that is, whether it is still needed by then).