Fixes: Remove faulty constness for QMap in DuiInputContext
RevBy: Pekka Vuorela
Details: The IC tests can crash because of faulty constness that was used when
filling the QMap, probably resulting in the wrong (that is, const) operator[]
overload to be selected. It turned out that most temporarily created variables
in DuiInputContext::getStateInformation() can be easily avoided, which was done
with this commit.