Fixes: NB#188429, Force input context activation on D-Bus reconnection
RevBy: Pekka Vuorela, Michael Hasselmann
Details: if setFocusWidget was called (due to user removing focus and setting
it back, for example) with D-Bus connection being down, that setFocusWidget()
call tried to activate and set active to true.  So, when the connection was
established again and setFocusWidget() was called, it didn't try to activate
at that point anymore, even though activation would now actually succeed.