Changes: estimate some surrounding text updates on IM side
RevBy: Pekka Vuorela
Details: this is done to support fix to NB#210004 (and other IM features)
that might need surrounding text to be up-to-date after sending, say, a
commit string but before the application has processed the input method event
resulting from that, IC has called updateWidgetInformation and that call has
been processed on the IM side.  The surrounding text (and consequently cursor
position too) is updated with the best guess that could be wrong -- there is
no way to know what the application might do when it gets a commit string,
for example.  Also, many cases such as having a selection at the time a
commit string is sent, are not supported.  You cannot rely on surrounding
text being correct due to this estimation, it only purports to help with
common cases.  Reliable way to get up-to-date state information will be
provided separately.