Fixes: NB#238370 - <Safety> Invalid read reported in meego-im-uiserver valgrind logs
RevBy: TrustMe
Details: When the underlying application goes away, there has been an
windowUnmapped-event that lead to invalid access to the remoteWindow member in
MImRotationAnimation. As there's a risk of similar cases arising in other clients
of the MImApplication signals and the mapped/unmapped events are not needed
anymore, let's remove them. While changing signals connected to
MImRotationAnimation the usage of contentUpdated signals was reduced
so that they are received only during orientation change.