Fixes: NB#227268 - [Performance] Virtual keyboard orientation switching is too slow
RevBy: Michael Hasselmann, Jan Arne Petersen
Details: Tracing shows that a considerable amount of time is spent in creating the
extra top level window for the rotation animation. Wrapping plugin widgets
into a proxy widget and chaining the MImRotationAnimation as a child
to MPassThruWindow significantly shortens the delay until the rotation plays.