Changes: Handle winid change in passthru window
Fixes: BMC#18772 - meego-im-uiserver is changing the window type after
it's window is mapped

Details: Handle changing of the win id and update the window type and
input region as soon as a valid win id is available. The X window is
created and assigned to after first call to QWidget::showFullScreen. So
the window type and input region needs to be set upon WinIdChange event.
In addition Qt overrides the _NET_WM_WINDOW_TYPE after it sends the
WinIdChangeEvent so it needs to be fixed again after the Show event
(which happens between creation of the X window and actual mapping).

RevBy: Michael Hasselmann