Changes: new plugin switching directions.
RevBy: Mohammad Anwari, Jan Arne Petersen, Chi Ning, Pekka Vuorela
Description: Either of SwitchPreparationBackward and SwitchPreparationForward
is given to the framework with switchPlugin function when the plugin
is preparing to switch to the previous/next plugin.
SwitchCanceled means that plugin decided to cancel preparations and not switch to
another one.