Add new preedit styles: unconvertible and active
RevBy: Kimmo Surakka, Pekka Vuorela

Add new styling enumerants for preedit.  "Unconvertible" indicates a
sub-region of the preedit which is non-interactive (cannot be
converted to real text due to UI limitations).  "Active" indicates a
preedit sub-region which is the subject of an active external UI
(e.g. a word choice list).

These are hooked up via default styling, but not into the
libmeegotouch style container abstraction.

Signed-off-by: Andy Ross <andy.ross@windriver.com>