Changes: Use base IC connection class, not glib-dbus in tests
RevBy: Jan Arne Petersen, Michael Hasselmann
Details: Since all the logic is now in the base class, we
no longer need to use the glib dbus connection class when testing.
For the sane reason we no longer need the stubs which were used to
implement the neccesary interface or stub out the dbus connection.