Fixes: Extension attributes are not registered after input context creation.
RevBy: Jon Nordby
This was because input context registers such extensions after receiving
a signal that connection to server is established. But input context is
created right after establishing such connection. To fix it, connecting
to server is deferred to mainloop. Also, this requirement is documented
in base class.

Regression introduced in 3780d15e1513bd032bdce2822b19f706b18cce73