New: Implement framework support for western numeric input override
RevBy: TrustMe
Partial fix of "NB#285309 - No way to input Western Arabic numerals (1234) when
inputMethodHint in textfield is set to Qt.ImhDigitsOnly and language to
Arabic".

Introduces a new IM query, Maliit::WesternNumericInputEnforcedQuery. A Qt
property using the key Maliit::InputMethodQuery::westernNumericInputEnforced
can be used as fallback, but it needs to be set on the focused widget itself
(applies for plain Qt apps, but not suitable for LMT apps, for example).