Changes: Remove custom QApplication class MIMApplication
RevBy: Jan Arne Petersen
Lets most code use the framework without needing a custom
QApplication subclass. The MImXApplication class is still
needed for correct server behavior under X11, but for
for QPA/Wayland the server can be run just with an QApplication.

Remaining dependencies on MImXApplication is now more explicit in
the code. Notable here is pixmap access for self-compositing.