New: SDK installed by default and include executable to get started
RevBy: Michael Hasselmann
The Maliit SDK, previously called Maliit Plugin SDK, is now installed
by default. This makes it clear which version of the framework
the SDK is for, and it no longer needs to be updated separately.

A maliit-sdk executable is also installed, and allows the
user to get started using the SDK. Run "maliit-sdk start"

Disabling the SDK can be done using CONFIG+=nosdk