Fixes: Building example applications stand-alone
RevBy: Michael Hasselmann
When stand-alone, examples are no longer in the source
tree and cannot refer to config.pri in the toplevel source
directory. Plugins do this directly, but applications never did.

Also no reason to install the applications, the user of the
examples can just run them directly from the directory.