Changes: Try to make test relying on dbus more reliable
RevBy: Jan Arne Petersen
Don't assume that the DBus message will be instantly dispatched,
so wait a bit before trying to process it.

When running on buildbot in a virtual X server,
Ut_MIMPluginManager::testDBusSetCalls() failed 4 out of 5 times.
With this change it succeeded 20 consecutive runs in the same
environment.

Not really a proper fix, as the test is still racy, but
at least much more reliable this way.