Fixes: NB#281340 - Often VKB rotation is not in sync.
RevBy: Jan Arne Petersen, Maksim Kirillov

Details:
Pixmap with application window snapshot is being created
in MInputContext and its handle is sent through d-bus
along with rotation angle.
This prevents snapshot being sometimes out of sync due to
the fact that pixmap in im-server was created to late
because of d-bus communication delay.
When im-server receives pixmap handle it uses it as a source
pixmap for rotation animation. Pixmap is then freed.