Changes: Guess an active subview from the list of enabled subviews, in case plugin was switched
RevBy: Viacheslav Sobolev, Antti Pulakka
It is possible to navigate between plugins without calling
MIMPluginManager::setActivePlugin, which leads to a situation where the active
plugin is asked for the active subview even though the framework never set one
on that plugin instance.

This commit is a workaround which checks active but empty subviews and then
tries to guess an active subview so that MImOnScreenPlugins::activeSubview
doesn't accidentially has to return an empty subview.