Add a surface API for use in plugins
Add a surface API which abstracts plugin windows and replaces the direct
use of a fullscreen QWidget hierarchy in plugins. With the new surface
API all the widget/window handling is done by the framework. This allows
to use keyboards also on non-compositing window managers or embeded in
applications.

RevBy: TrustMe