Don't use qtAddLibrary() in maliit-plugins.prf
RevBy: TrustMe
This macro adds a trailing 'd' to the name of DLL
files on Windows when building in Debug mode.
This does not match how we name the DLL when installing,
causing maliit plugins to fail to link.
The new way is the same as for the .pc file.