Use signals instead of callbacks for Gtk+ DBus connection
Use proper GObject signals on MeegoIMContextDbusObj Instead of using
meego_imcontext_client_ callbacks which need to be implemented in
libraries linking to the static connection libraries.

Allows to convert that static library into a proper connection-glib and
use it directly by maliit-glib instead of routing the calls through the
input context.

For more cleanup to be done in the Gtk+ input context see
https://bugs.maliit.org/show_bug.cgi?id=156

RevBy: Krzesimir Nowak, Michael Hasselmann