Find gtk-query-immodules-x script on Ubuntu 12.x and higher
RevBy: Jan Arne Petersen
The symlink in /usr/bin is gone (for gtk-query-immodules-2.0), at least on
Ubuntu 12.04, and we still need to call the script manually for developers
using Maliit from the git repo (or tarballs). Also added the workaround for
GTK+3, so we don't rely on the symlink there.