Explicitly specify temporary/persistent settings for MImSettings
RevBy: NeedsRev
Not specifying what to use is a programmer error and will abort.
This fixes "make check" overwriting the normal settings, and
ensures that such an issue does not sneak in again.

Only maliit-server standalone or Maliit in server-hosted mode
will use persistent settings. Tests and application hosted will
use temporary.
Because we don't want MImSettings to be public API, a wrapper
is added to MImServer for use in server-hosted mode.