Fixes: MALIIT#188 - maliit-server forgets active subview
RevBy: Krzesimir Nowak
The goal is to guess an active subview from the enabled list, if none was set,
and to insert the currently active subview into enabled list.

Removed the unfortunate side effect in
MImOnScreenPlugins::updateEnabledSubviews that would also set the active
subview. Then fixed MImOnScreenPlugins::updateActiveSubviews by making it
single-exit, except for the trivial case when active subview didn't change,
which helps setting the active subview correctly.

In related news: The subview handling code in the framework is seriously
complex.