Fixes: MALIIT#197 - Read initial SW_TABLET_MODE state correctly
We found a small bug in the SW_TABLET_MODE support: the tablet mode
state is not read correctly during startup.

The ioctl request fired here is invalid and always returns error - its
a silly coding mistake.

This means that we never accurately initialise the tablet switch state
(e.g. to figure out if the system was in tablet mode at boot-time).

Firing the correct ioctl solves the issue. Fixes
http://bugs.sugarlabs.org/ticket/4051

RevBy: Michael Hasselmann