Make input mask work for Qt5 and the xcb backend
With Qt5, we cannot use the old X11 API but have to use the (fairly similar)
xcb API instead. The idea is the still the same: A "fullscreen" surface needs
to punch a hole in its (mostly translucent) window so that input events can be
passed through to the application window.

RevBy: Michael Hasselmann