Improved default configuration handling
Previously, maliit was doing a certain amount of automatic config
manipulation (mostly based on compile-time options) and writing this
out to disk.

This was not ideal, because the compile-time preferences might change
during a software update, but the already-written user config would
override that. Also, followup commits will make the automatic detection
of active/enabled subviews a lot more dynamic in some cases.

We solve those issues by not writing out the config changes to disk
when they were driven by internal auto-detection.

MImOnScreenPlugins now presents an "empty" subview when no configuration
is present, and MImPluginManager acts on that by auto-activating the first
enabled subview in the configuration (if present), otherwise falling back
to the first available subview.

MALIIT_DEFAULT_SUBVIEW is removed in the process; the first available
subview from a plugin will be used by default when none is specified
in the configuration. Distributors can override this by shipping a
configuration in /etc/xdg if desired.

guessActiveSubview is removed because there are currently no codepaths that
can hit the condition that originally justfieid it.

The MImOnScreenPlugins tests needed updating now that an empty config
presents an empty subview by default; similar tests (and tests for the
new logic) have been placed in ImPluginManager tests.

RevBy: Michael Hasselmann