Auto-detect subviews based on system locale
When no user configuration is provided, we now attempt to automatically
available suitable subviews based on the system locale information.

This is mainly targetted at the keyboard plugin; the automatic detection
will need extending if we want to consider other plugins in future
which don't base their subviews around locale codes.

In this case, we use QLocale.uiLanguages() to determine which subviews
to enable, and in which order. Internally, this uses the LANGUAGE variable
(a priority-ordered list of preferred languages):
http://www.gnu.org/software/gettext/manual/html_node/The-LANGUAGE-variable.html
and if LANGUAGE is not supplied it falls back on the normal single-entry
LANG/LC_MESSAGES variables.

When combined with the behaviour of automatically enabling the first subview
when no configuration is supplied, this results in a sensible choice of
default keyboard layout, and a sensible choice of alternate languages
available through the language key.

A side effect of this change is that a configuration-supplied active subview
will no longer be automatically added to the enabled subviews list. This
is intentional: think about the case where the user runs their system
in French, with German as a fallback. They use the language key to switch
the OSK to German, so now the active keyboard (German) is written to
the config, with enabled still not present in the config (i.e. it will
continue to be auto-detected).

Now the impressively-multilingual user makes a drastic change, and modifies
the system locale to be Spanish with a fallback on Russian. In this case,
we do not want the saved active keyboard layout (German) to take effect,
the user has made no explicit attempt to configure Maliit so we want to
continue autodetecting everything, even in the face of underlying
system configuration changes. This is why active keyboard configurations
do not get automatically enabled any more.

This means that settings applications must ensure that a subview is enabled
before trying to activate it, which seems logical anyway. The settings
example app has been modified accordingly.

RevBy: Michael Hasselmann