Move from std::tr1::shared_ptr to QSharedPointer.
RevBy: Jan Arne Petersen

Using two shared pointer implementations is messy at best. The only
inconvenience of moving to QSharedPointer is that it does not have a
reset(T* ptr) method. This in turn forces us to use awkward constructs
for conditional construction of shared pointer:

QSharedPointer<Foo> f;
if (a)
    f = QSharedPointer<Foo>(new A(p1, p2, ...));
else if
    f = QSharedPointer<Foo>(new B(p1, p2, ...));

Instead we move the construction to function returning a shared
pointer.