Implement show/hide transition support for QML plugins
New property MInputMethodQuick.active qml plugins can track to show and
hide UI. Previously this was done by the framework.

Forward-ported from 0.81. There this was opt-in behavior, but on
unstable 0.9x we can just switch to such.

Note: hiding currently instantly hides the surface so transition will
not be visible. Reported as MALIIT#205.

RevBy: Michael Hasselmann