Make MInputMethodQuickPlugin standalone.
RevBy: TBD

MInputMethodQuickPlugin was initially designed to be derived by plugins, so it
would provide an implementation of qmlFileName() and name() vfunc. Right now
this idea was discarded in favour of quick plugin factory. So now we don't need
the qmlFileName vfunc and instead we pass the QML file to be loaded in plugins'
constructor.

That is - remove the qmlFileName vfunc, implement