Fixes: Windows symbol import/export in dll
On windows symbols are normally NOT automatically exported in DLLs,
leading to missing .lib files and failure to link.

CMake has an option to automatically export all symbols of shared
libraries by generating .def files. This option is now enabled in
CMakeLists.txt.

Furthermore GenerateExportHeader is included in CMakelists and
used to generate an export header for maliit-plugins which in
turn is used to export MAbstractInputMethod. This header is also
added to install.