Re-show the keyboard on wayland surrounding text changes
If we the keyboard is hidden (using the close gesture or a similar
method) and the surrounding text changes (suggesting the user tapped the
same text field but in a different place), re-show the keyboard.

This assumes the user hid the keyboard manually, then tapped the
textfield again to request it be reshown.