Support resizing the input panel when using Wayland
This relies on a new "configure" event that has been added to the
input-panel-surface-v1 protocol, which allows the compositor to request
a new size for the input panel. This is primarily intended to allow the
compositor to position the panel in such a way that it avoids things
like task managers.