Fix application orientation angle back to clockwise
Got changed in commit b9f448fcc

Slightly hard to follow, but the Qt::LandscapeOrientation & friends
are defined with clockwise orientation. The QScreen::angleBetween()
doesn't explicitly define the direction but the results match how much
the physical screen would need to be rotated clockwise, and commonly
clockwise angles are used within Qt.

But now, if the display is rotated 90 degrees clockwise, the application
UI should be rotated to opposite direction to remain shown correctly,
i.e. to 270.

Thus switch the parameter order to angleBetween() and define the resulting
appOrientation angle to hopefully avoid confusion in the future.